Refactor worker
