--- standalone/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "standalone-skywalking-worker"
-version = "0.8.0-dev"
-authors = ["Apache Software Foundation", "jmjoy <jmjoy@apache.org>", "Yanlong He <heyanlong@apache.org>"]
-description = "Apache SkyWalking PHP Standalone worker."
-edition = "2021"
-rust-version = "1.65"
-repository = "https://github.com/apache/skywalking-php"
-license = "Apache-2.0"
-readme = "README.md"
-publish = false
-
-[dependencies]
-anyhow = { version = "1.0.72", features = ["backtrace"] }
-clap = { version = "=4.3.24", features = ["derive"] }
-clap_lex = "=0.5.0"
-skywalking-php = { path = ".." }
-tracing = { version = "0.1.37", features = ["attributes"] }
-tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
