+++ worker/Cargo.toml
@@ -13,26 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-[workspace]
-members = [
-    ".",
-    "scripts",
-    "worker",
-]
-
-[workspace.package]
-version = "0.8.0-dev"
-authors = ["Apache Software Foundation", "jmjoy <jmjoy@apache.org>", "Yanlong He <heyanlong@apache.org>"]
-edition = "2021"
-rust-version = "1.65"
-repository = "https://github.com/apache/skywalking-php"
-license = "Apache-2.0"
-
 [package]
-name = "skywalking-php"
+name = "skywalking-php-worker"
 version = { workspace = true }
 authors = { workspace = true }
-description = "Apache SkyWalking PHP Agent."
+description = "Apache SkyWalking PHP Worker."
 edition = { workspace = true }
 rust-version = { workspace = true }
 repository = { workspace = true }
@@ -40,39 +25,27 @@ license = { workspace = true }
 readme = "README.md"
 publish = false
 
-[lib]
-name = "skywalking_agent"
-crate-type = ["lib", "cdylib"]
-
 [features]
-kafka-reporter = ["skywalking-php-worker/kafka-reporter"]
+standalone = ["clap", "clap_lex", "tracing-subscriber"]
+kafka-reporter = ["skywalking/kafka-reporter", "rdkafka/sasl"]
 
 [dependencies]
 anyhow = { version = "1.0.72", features = ["backtrace"] }
 bincode = "1.3.3"
-dashmap = "5.5.0"
-futures-util = "0.3.28"
-hostname = "0.3.1"
+clap = { version = "=4.3.24", features = ["derive"], optional = true }
+clap_lex = { version = "=0.5.0", optional = true }
 libc = "0.2.147"
 once_cell = "1.18.0"
-phper = "0.12.0"
 prost = "0.11.9"
 rdkafka = { version = "0.32.2", optional = true }
+serde_json = { version = "1.0.104", features = ["preserve_order"] }
 skywalking = { version = "0.8.0", features = ["management"] }
-skywalking-php-worker = { path = "worker" }
-systemstat = "0.2.3"
-thiserror = "1.0.44"
 tokio = { version = "1.29.1", features = ["full"] }
 tokio-stream = "0.1.14"
 tonic = { version = "0.8.3", features = ["tls", "tls-roots"] }
 tracing = { version = "0.1.37", features = ["attributes"] }
-tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
-url = "2.4.0"
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"], optional = true }
 
-[dev-dependencies]
-axum = "0.6.19"
-fastcgi-client = "0.9.0"
-reqwest = { version = "0.11.18", features = ["trust-dns", "json", "stream"] }
-
-[build-dependencies]
-phper-build = "0.12.0"
+# [[bin]]
+# name = "skywalking-php-worker"
+# required-features = ["standalone", "kafka-reporter"]
